# Cache socket.local_address and socket.remote_address