--- a/src/crystal/system/unix/socket.cr
+++ b/src/crystal/system/unix/socket.cr
@@ -239,28 +239,35 @@ module Crystal::System::Socket
     end
   end
 
+  @local_address_cache : ::Socket::IPAddress?
+  @remote_address_cache : ::Socket::IPAddress?
+
   private def system_local_address
-    sockaddr6 = uninitialized LibC::SockaddrIn6
-    sockaddr = pointerof(sockaddr6).as(LibC::Sockaddr*)
-    addrlen = sizeof(LibC::SockaddrIn6).to_u32!
+    @local_address_cache ||= begin
+      sockaddr6 = uninitialized LibC::SockaddrIn6
+      sockaddr = pointerof(sockaddr6).as(LibC::Sockaddr*)
+      addrlen = sizeof(LibC::SockaddrIn6).to_u32!
 
-    if LibC.getsockname(fd, sockaddr, pointerof(addrlen)) != 0
-      raise ::Socket::Error.from_errno("getsockname")
+      if LibC.getsockname(fd, sockaddr, pointerof(addrlen)) != 0
+        raise ::Socket::Error.from_errno("getsockname")
+      end
+
+      ::Socket::IPAddress.from(sockaddr, addrlen)
     end
-
-    ::Socket::IPAddress.from(sockaddr, addrlen)
   end
 
   private def system_remote_address
-    sockaddr6 = uninitialized LibC::SockaddrIn6
-    sockaddr = pointerof(sockaddr6).as(LibC::Sockaddr*)
-    addrlen = sizeof(LibC::SockaddrIn6).to_u32!
+    @remote_address_cache ||= begin
+      sockaddr6 = uninitialized LibC::SockaddrIn6
+      sockaddr = pointerof(sockaddr6).as(LibC::Sockaddr*)
+      addrlen = sizeof(LibC::SockaddrIn6).to_u32!
 
-    if LibC.getpeername(fd, sockaddr, pointerof(addrlen)) != 0
-      raise ::Socket::Error.from_errno("getpeername")
+      if LibC.getpeername(fd, sockaddr, pointerof(addrlen)) != 0
+        raise ::Socket::Error.from_errno("getpeername")
+      end
+
+      ::Socket::IPAddress.from(sockaddr, addrlen)
     end
-
-    ::Socket::IPAddress.from(sockaddr, addrlen)
   end
 
   {% if flag?(:openbsd) %}