--- a/src/socket/ip_socket.cr
+++ b/src/socket/ip_socket.cr
@@ -1,11 +1,7 @@
 class IPSocket < Socket
   # Returns the `IPAddress` for the local end of the IP socket.
-  def local_address
-    system_local_address
-  end
+  getter local_address : Socket::IPAddress { system_local_address }
 
   # Returns the `IPAddress` for the remote end of the IP socket.
-  def remote_address
-    system_remote_address
-  end
+  getter remote_address : Socket::IPAddress { system_remote_address }
 end