`HTTP::Request.from_io` queries the io's local_address and remote_address which involves calling `getsockname` and `getpeername` into the system on every request. Doing that can be expensive.

Here I propose caching this information with the socket as it won't ever change. This improves the requests/second throughput.

Before:

```
$ wrk -t12 -c100 -d30s http://127.0.0.1:8080
Running 30s test @ http://127.0.0.1:8080
  12 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.95ms  288.83us  12.16ms   86.64%
    Req/Sec     8.48k   523.58    20.32k    92.73%
  3042411 requests in 30.10s, 293.05MB read
Requests/sec: 101077.24
Transfer/sec:      9.74MB
```

After: (~9% improvement)
```
$ wrk -t12 -c100 -d30s http://127.0.0.1:8080
Running 30s test @ http://127.0.0.1:8080
  12 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.89ms  488.80us  25.49ms   94.96%
    Req/Sec     9.20k   377.40    10.80k    80.03%
  3296845 requests in 30.01s, 317.56MB read
Requests/sec: 109868.71
Transfer/sec:     10.58MB
```