@straight-shoota 2021-05-30
This looks like a good idea. In fact, I've already considered this, too. In the win32 implementation, `AcceptEx` gives both socket addresses, and they could be stored without ever having to query the values (on accepted sockets).

However, I believe the proposed implementation is inappropriate. There should be no modification to the system-specific methods. Instead, the values should be memoized in the public-facing API.