@lbguilherme 2021-05-30
@straight-shoota Initially I was planning to change Socket directly. But then Windows' `AcceptEx` can return the address and this can be cached right way. If this isn't an implementation detail, how will the Windows port be able to do that? Having two layers of variables doesn't seems right.