@lbguilherme 2021-05-30
@straight-shoota I modified to cache at the outer layer. Please take a look.