@straight-shoota 2021-05-31
There need to be resets in `#connect` and `#bind`. Those are the methods that affect the address values. Rebinding probably doesn't work in practice. But connecting to a different address definitely works, at least for datagram sockets.

A spec to test with re-connecting could look like this:
```cr
client = UDPSocket.new
client.connect "127.0.0.1", 1
client.remote_address.port.should eq 1
client.connect "127.0.0.1", 2
client.remote_address.port.should eq 2
```

A couple more specs probably wouldn't hurt either.