@lbguilherme 2021-06-01
@straight-shoota I added a reset after `connect` and `bind`, and your spec. I couldn't find any other case where it would reset beside `connect`. Rebinding doesn't work, but I left it there anyway.