@straight-shoota 2021-06-04
The test is failing for IPv6: `Hostname lookup for ::1 failed: Address family for hostname not supported`.
This could be a limitation of the CI environment? Does it succeed locally?