@lbguilherme 2021-06-04
I didn't notice it before because my local env doesn't support IPv6 (I work inside a container most of the time).

Turns out that this new spec already fails for IPv6 against master. It seems like you can't "connect" a UDP socket to an IPv6 address. Something is going wrong with name resolution.

For now I moved the spec around to only run with IPv4.